# release-19.2: sql/pgwire: avoid new memory allocation in writeTextDatum(Date)